project_update

+++ src/main/java/com/project/demo/controller/StockController.java
@@ -14,11 +14,13 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 
 import com.project.demo.dao.BookRepository;
 import com.project.demo.dao.StockRepository;
 import com.project.demo.models.Book;
+import com.project.demo.models.Manager;
 import com.project.demo.models.Stock;
 
 @Controller
@@ -45,8 +47,27 @@ public class StockController {
 	@Autowired
 	BookRepository booksrepo;
 	
+	@GetMapping("/addstock")
+	public String AddStock()
+	{
+		return "AddStock";
+	}
+	
+	@GetMapping("/updatestock")
+	public String UpdateStock()
+	{
+		return "UpdateStock";
+	}
+	
+	@GetMapping("/deletestock")
+	public String DeleteStock()
+	{
+		return "DeleteStock";
+	}
+	
+	
 	@PostMapping("/addstock")
-	public String AddStock(@RequestBody Stock stock,Model model, HttpSession session)
+	public String AddStock(@ModelAttribute("stock") Stock stock,Model model, HttpSession session)
 	{
 		if(!is_manager(session))
 		{
@@ -61,17 +82,24 @@ public class StockController {
 			return "home";
 		}
 		
-		Stock savedStock =  stockrepo.saveStock(stock);
-		int cnt = savedStock.getBooksavailable();
+		int flag =  stockrepo.saveStock(stock);
+		
+		if(flag!=1)
+		{
+			model.addAttribute("error","SOMETHING WENT WRONG, PLEASE TRY AGAIN");
+			return "home";
+		}
+		
+		int cnt = stock.getBooksavailable();
 		for(int i=1;i<=cnt;i++)
 		{
 			Book book = new Book();
-			book.setShelfId(101);
+			book.setShelfid(101);
 			book.setLanguage("English");
-			book.setPublications(savedStock.getPublications());
-			book.setAuthor(savedStock.getAuthor());
-			book.setTitle(savedStock.getTitle());
-			int flag = booksrepo.saveBook(book);
+			book.setPublications(stock.getPublications());
+			book.setAuthor(stock.getAuthor());
+			book.setTitle(stock.getTitle());
+			int done = booksrepo.saveBook(book);
 		}
 		
 		model.addAttribute("error","STOCK ADDED SUCCESSFULLY");
@@ -79,32 +107,56 @@ public class StockController {
 	}
 	
 	@GetMapping("/getstocks")
-	public @ResponseBody String GetStock(Model model)
+	public String GetStock(Model model,HttpSession session)
 	{
+		if(is_student(session) || is_staff(session))
+		{
+			model.addAttribute("error","NOT AUTHORISED TO SEE STOCK DETAILS");
+			//return "NOT AUTHORISED TO SEE RETAILER DETAILS";
+			return "home";
+		}
+		
+		if(!is_manager(session))
+		{
+			model.addAttribute("error","LOGIN AS A MANAGER OF THE BOOKS DEPARTMENT BEFORE VIEWING THIS PAGE");
+			//return "LOGIN AS A MANAGER OF THE RETAILER DEPARTMENT BEFORE VIEWING THIS PAGE";
+			return "Login";		
+		}
+		
+		Manager manager = (Manager)session.getAttribute("manager");
+		int cur_id = manager.getDeptid();
+		
+		if(cur_id!=101)
+		{
+			model.addAttribute("error","LOGIN AS A MANAGER OF THE BOOKS DEPARTMENT BEFORE VIEWING THIS PAGE");
+			//return "LOGIN AS A MANAGER OF THE RETAILER DEPARTMENT BEFORE VIEWING THIS PAGE";
+			return "Login";
+		}
+		
 		List <Stock> list = stockrepo.getAllStocks();
 		model.addAttribute("qresult",list);
 		return "Stocks";
 	}
 	
-	@GetMapping("/getstock/{title}/{author}/{publications}")
+	@GetMapping("/getstock")
 	public String GetStock(
-			@PathVariable("title") String title, 
-			@PathVariable("author") String author, 
-			@PathVariable("publications") String publications,
+			@RequestParam("title") String title, 
+			@RequestParam("author") String author, 
+			@RequestParam("publications") String publications,
 			Model model, HttpSession session
 			)
 	{
 		
 		 Stock stock =  stockrepo.getStock(title, author, publications);
-		 model.addAttribute("stock",stock);
+		 model.addAttribute("q",stock);
 		 return "Stock";
 	}
 	
-	@DeleteMapping("/deletestock/{title}/{author}/{publications}")
-	public @ResponseBody String DeleteStock(
-			@PathVariable("title") String title, 
-			@PathVariable("author") String author, 
-			@PathVariable("publications") String publications,
+	@PostMapping("/deletestock")
+	public String DeleteStock(
+			@RequestParam("title") String title, 
+			@RequestParam("author") String author, 
+			@RequestParam("publications") String publications,
 			Model model,HttpSession session
 			)
 	{
@@ -121,6 +173,14 @@ public class StockController {
 			return "home";
 		}
 		
+		Stock stock = stockrepo.getStock(title, author, publications);
+		
+		if(stock.getBooksavailable()>0 || stock.getBooksissued()>0)
+		{
+			model.addAttribute("error","CAN'T DELETE THIS STOCK AS SOME BOOKS ARE STILL THERE IN THE LIBRARY");
+			return "home";
+		}
+		
 		int flag = stockrepo.deleteStockById(title, author, publications);
 		if(flag!=1)
 		{
@@ -132,26 +192,31 @@ public class StockController {
 		return "home";
 	}
 	
-	@PutMapping("/updatestock")
-	public @ResponseBody String UpdateStock(@RequestBody Stock stock,Model model, HttpSession session)
+	@PostMapping("/updatestock")
+	public String UpdateStock(@ModelAttribute("stock") Stock stock,Model model, HttpSession session)
 	{
 		if(!is_manager(session))
 		{
 			model.addAttribute("error","YOU DON'T HAVE APPROPRIATE PERMISSIONS");
-			return "YOU DON'T HAVE APPROPRIATE PERMISSIONS";
-			//return "home";
+			//return "YOU DON'T HAVE APPROPRIATE PERMISSIONS";
+			return "home";
 		}
 		
 		int dept = (int)session.getAttribute("Dept");
 		if(dept!=101)
 		{
 			model.addAttribute("error","ONLY MANAGERS OF THE BOOKS DEPARTMENT CAN ADD BOOKS");
-			return "ONLY MANAGERS OF THE BOOKS DEPARTMENT CAN ADD BOOKS";
-			//return "home";
+			//return "ONLY MANAGERS OF THE BOOKS DEPARTMENT CAN ADD BOOKS";
+			return "home";
+		}
+		Stock intialStock = stockrepo.getStock(stock.getTitle(),stock.getAuthor(),stock.getPublications());
+		if(intialStock.getBooksavailable()==-1)
+		{
+			model.addAttribute("error","STOCK DETAILS DOESN'T EXIST");
+			return "home";
+			
 		}
 		
-		// for now assume that the stock is being added
-		Stock intialStock = stockrepo.getStock(stock.getTitle(),stock.getAuthor(),stock.getPublications());
 		int prev = intialStock.getBooksavailable();
 		Stock updatedStock =  stockrepo.updateStock(stock);
 		int cur = updatedStock.getBooksavailable();
@@ -159,7 +224,7 @@ public class StockController {
 		for(int i=1;i<=cur-prev;i++)
 		{
 			Book book = new Book();
-			book.setShelfId(101);
+			book.setShelfid(101);
 			book.setLanguage("English");
 			book.setPublications(updatedStock.getPublications());
 			book.setAuthor(updatedStock.getAuthor());
@@ -168,7 +233,6 @@ public class StockController {
 		}
 		
 		model.addAttribute("error","STOCK UPDATED SUCCESSFULLY");
-		return "STOCK UPDATED SUCCESSFULLY";
-		//return "home";
+		return "home";
 	}
 }
